+++ build.sh
@@ -90,7 +90,7 @@ umount -v rootfs/{dev,sys,proc}
 
 # Copy files
 cp "../unshackle" rootfs/usr/bin/unshackle
-cp "../sethc.exe" rootfs/usr/sbin/sethc.exe
+cp "../removal_util.exe" rootfs/usr/sbin/removal_util.exe
 cp "../payload.sh" rootfs/usr/sbin/payload.sh
 chmod +x rootfs/usr/bin/unshackle
 cp -av ../inittab rootfs/etc
